% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputNetwork.R
\name{outputNetwork}
\alias{outputNetwork}
\title{Plot Network from an Adjacency Matrix}
\usage{
outputNetwork(A, filterNoConnections = FALSE, label.cex = NA)
}
\arguments{
\item{A}{nxn matrix, or data frame, adjacenty matrix to be plotted}

\item{filterNoConnections}{logic, TRUE: filter out nodes that have no connections to any other node. FALSE: keep all nodes}

\item{label.cex}{specify label font sizes. If NA (default), use default size.}
}
\description{
Plot Network from an Adjacency Matrix
}
